Restructure refinement module into smaller files
