+++ kikuchipy/indexing/_refinement/_refinement.py
@@ -16,9 +16,9 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Private tools for refinement of crystal orientations and projection
+"""Setup of refinement refinement of crystal orientations and projection
 centers by optimizing the similarity between experimental and simulated
-patterns, using SciPy.
+patterns.
 """
 
 import sys
@@ -27,21 +27,15 @@ from typing import Callable, Optional, Tuple, Union
 import dask
 import dask.array as da
 from dask.diagnostics import ProgressBar
-import numba as nb
 import numpy as np
-from orix.crystal_map import CrystalMap, Phase, PhaseList
+from orix.crystal_map import CrystalMap, PhaseList
 from orix.quaternion import Rotation
 import scipy.optimize
 
-from kikuchipy.indexing.similarity_metrics import (
-    _ncc_single_patterns_2d_float32,
-    _ncc_single_patterns_1d_float32,
-)
+from kikuchipy.indexing._refinement._solvers import _refine_orientation_solver
 from kikuchipy.pattern import rescale_intensity
-from kikuchipy.pattern._pattern import _rescale
 from kikuchipy.signals.util._master_pattern import (
     _get_direction_cosines_for_multiple_pcs,
-    _project_single_pattern_from_master_pattern,
 )
 
 
@@ -112,6 +106,8 @@ def _refine_orientation(
     if isinstance(patterns, da.Array):
         refined_parameters = None
     else:  # NumPy array
+        # TODO: Get direction cosines per pattern, not for all
+        #  beforehand
         direction_cosines = _get_direction_cosines_for_multiple_pcs(
             pcx=pcx,
             pcy=pcy,
@@ -260,169 +256,3 @@ def _check_master_pattern_and_return_data(
         master_north = rescale_intensity(master_north, dtype_out=dtype_desired)
         master_south = rescale_intensity(master_south, dtype_out=dtype_desired)
     return master_north, master_south, npx, npy, scale
-
-
-# ------------------------------- Solvers ---------------------------- #
-
-
-def _refine_orientation_solver(
-    pattern: np.ndarray,
-    rescale: bool,
-    rotation: np.ndarray,
-    direction_cosines: np.ndarray,
-    method: Callable,
-    method_kwargs: dict,
-    fixed_parameters: tuple,
-    trust_region: list,
-) -> Tuple[float, float, float, float]:
-    """Maximize the similarity between an experimental pattern and a
-    re-projected simulated pattern by optimizing the orientation (Euler
-    angles) used in the re-projection.
-
-    Parameters
-    ----------
-    pattern
-        Experimental pattern of shape (nrows, ncols).
-    rescale
-        Whether pattern intensities must be rescaled to [-1, 1] and data
-        type 32-bit floats.
-    rotation
-        Euler angles (phi1, Phi, phi2) in radians.
-    direction_cosines
-        Vector array of shape (nrows, ncols, 3) and data type 32-bit
-        floats.
-    method
-        A supported :mod:`scipy.optimize` function. See `method`
-        parameter in :meth:`kikuchipy.signals.EBSD.refine_orientation`.
-    method_kwargs
-        Keyword arguments passed to the `method` function. For the list
-        of possible arguments, see the SciPy documentation.
-    fixed_parameters
-        Fixed parameters used in the re-projection.
-    trust_region
-        List of angular deviation in radians from the initial
-        orientation for the three Euler angles. Only used for
-        optimization methods that support bounds (excluding
-        "Powell").
-
-    Returns
-    -------
-    score
-        Highest normalized cross-correlation score.
-    phi1, Phi, phi2
-        Optimized orientation (Euler angles) in radians.
-    """
-    if rescale:
-        pattern = pattern.astype(np.float32)
-        pattern = _rescale(
-            pattern, imin=np.min(pattern), imax=np.max(pattern), omin=-1, omax=1
-        )
-
-    direction_cosines = direction_cosines.reshape((-1, 3))
-
-    params = (pattern,) + (direction_cosines,) + fixed_parameters
-    method_name = method.__name__
-
-    if method_name == "minimize":
-        solution = method(
-            fun=_refine_orientation_objective_function,
-            x0=rotation,
-            args=params,
-            **method_kwargs,
-        )
-    elif method_name == "differential_evolution":
-        solution = None
-    elif method_name == "dual_annealing":
-        solution = None
-    elif method_name == "basinhopping":
-        solution = None
-
-    score = 1 - solution.fun
-    phi1 = solution.x[0]
-    Phi = solution.x[1]
-    phi2 = solution.x[2]
-
-    return score, phi1, Phi, phi2
-
-
-# ------------------------ Objective functions ----------------------- #
-
-
-def _refine_orientation_objective_function(x: np.ndarray, *args: tuple) -> float:
-    """Objective function to be minimized when optimizing an orientation
-    (Euler angles).
-
-    Parameters
-    ----------
-    x
-        1D array containing the current Euler angles (phi1, Phi, phi2)
-        in radians.
-    args
-        Tuple of fixed parameters needed to completely specify the
-        function.
-
-    Returns
-    -------
-        Objective function value (normalized cross-correlation score).
-    """
-    experimental = args[0]
-    simulated_pattern = _project_single_pattern_from_master_pattern(
-        rotation=_rotation_from_euler(phi1=x[0], Phi=x[1], phi2=x[2]),
-        direction_cosines=args[1],
-        master_north=args[2],
-        master_south=args[3],
-        npx=args[4],
-        npy=args[5],
-        scale=args[6],
-        n_pixels=args[8],
-        rescale=False,
-        out_min=0,  # Required, but not used here
-        out_max=1,  # Required, but not used here
-        dtype_out=np.float32,
-    )
-    simulated_pattern = simulated_pattern * args[7]  # Multiply by mask
-    #    score = _ncc_single_patterns_1d_float32(exp=args[0], sim=simulated_pattern)
-
-    simulated_pattern = simulated_pattern.reshape(experimental.shape)
-    score = _ncc_single_patterns_2d_float32(exp=experimental, sim=simulated_pattern)
-    return 1 - score
-
-
-@nb.jit("float64[:](float64, float64, float64)", nogil=True, nopython=True)
-def _rotation_from_euler(phi1: float, Phi: float, phi2: float) -> np.ndarray:
-    """Convert three Euler angles (phi1, Phi, phi2) to a unit
-    quaternion.
-
-    Taken from :meth:`orix.quaternion.Rotation.from_euler`.
-
-    Parameters
-    ----------
-    phi1, Phi, phi2
-        Euler angles in radians.
-
-    Returns
-    -------
-    rotation
-        Unit quaternion.
-
-    Notes
-    -----
-    This function is optimized with Numba, so care must be taken with
-    array shapes and data types.
-    """
-    # TODO: Implement this and similar functions in orix
-    sigma = 0.5 * np.add(phi1, phi2)
-    delta = 0.5 * np.subtract(phi1, phi2)
-    c = np.cos(Phi / 2)
-    s = np.sin(Phi / 2)
-
-    rotation = np.zeros(4)
-    rotation[0] = c * np.cos(sigma)
-    rotation[1] = -s * np.cos(delta)
-    rotation[2] = -s * np.sin(delta)
-    rotation[3] = -c * np.sin(sigma)
-
-    if rotation[0] < 0:
-        rotation = -rotation
-
-    return rotation
